FIX: Corregir los errores de compilación de los mappers y arreglo de la antigua lógica para que pueda correr la aplicación - Se realizó una limpieza total del esquema de la base de datos para sincronizar Flyway e Hibernate. - Corrección en EntrenadorRepository: se renombró findByUsuarioEmail a findByUsuario_Email para permitir la navegación correcta de propiedades en Spring Data JPA. - Implementación de @JdbcTypeCode(SqlTypes.NAMED_ENUM) en las entidades Usuario y Entrenador para el mapeo correcto de tipos ENUM nativos de PostgreSQL (usuario_estado, usuario_rol y estado_revision_tipo). - Se aseguró la compatibilidad entre los Enums de Java y las definiciones de tipos en los scripts de Flyway.

+++ backend/src/main/java/grit/sistema/backend/dto/atleta/AtletaRequestDTO.java
@@ -6,6 +6,7 @@ import grit.sistema.backend.model.enums.TipoServicio;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record AtletaRequestDTO(
@@ -32,7 +33,7 @@ public record AtletaRequestDTO(
         String genero,
 
         @DecimalMin(value = "30.0", message = "El peso mínimo es 30.0 kg")
-        Double pesoKg,
+        BigDecimal pesoKg,
 
         @Min(value = 100, message = "La altura mínima es 100 cm")
         Integer alturaCm,
